fixed google auth
role required a default value, for now the default value is "none" but we need to implement something that lets the user choose its role. It can either be something we do right after better-auth creates account, or once the user reaches the dashboard, if they have "none" role we make them choose one. I prefer the first option.